Add feedback server to server
